Create billsplitter.py
I completed my first Python project from Hyerskill - the JetBrains Academy.

This is a mini project in which the user can register some friends, including himself, for a party and then split the bill among themselves.

There is also a lucky feature which the user can enable and one of the friends won't have to pay any money, and his part will be d…